Update Radar Leumit.py

+++ pages/Radar Leumit.py
@@ -41,10 +41,11 @@ selected_position_group = st.sidebar.selectbox(
 min_minutes_played = st.sidebar.slider(
     "Filter by Minimum Minutes Played:",
     min_value=0,
-    max_value=int(df['Minutes played'].max()),  # Convert max_value to int
+    max_value=int(df['Minutes played'].astype(int).max()),  # Convert both min and max values to int
     step=1,
     value=0
 )
+
 
 # Filter the DataFrame based on the selected position group and minimum minutes played
 filtered_players = df[df["Position"].str.contains("|".join(position_mapping[selected_position_group]), case=False) & (df['Minutes played'] >= min_minutes_played)]
